chg: usr: treasures are prioritized in a conf file
