+++ conf/system/treasures.json
@@ -1,0 +1,13 @@
+{
+    "blessing_of_the_moon": 2,
+    "celestial_alignment": 10,
+    "elementary_studies": 3,
+    "fire_staff": 5,
+    "fireball_volley": 4,
+    "improved_lightning_bolt": 6,
+    "nature_staff": 8,
+    "positive_equilibrium": 7,
+    "ring_of_protection": 9,
+    "tempest_fury": 1,
+    "treasure_chest": 11
+}
