+++ modules/settings/__init__.py
@@ -9,6 +9,7 @@ from .conf import (
     mercsAbilities,
     ability_order,
     settings_dict,
+    treasures_priority,
 )
 
 
@@ -19,4 +20,5 @@ __all__ = [
     "mercsAbilities",
     "ability_order",
     "settings_dict",
+    "treasures_priority",
 ]
